+++ core/src/main/java/org/wildfly/extras/creaper/core/ServerType.java
@@ -1,8 +1,10 @@
 package org.wildfly.extras.creaper.core;
 
 /**
- * AS7 or WildFly.
+ * @deprecated use {@link org.wildfly.extras.creaper.core.online.ManagementProtocol} instead,
+ * this will be removed before 1.0
  */
+@Deprecated
 public enum ServerType {
     AS7,
     WILDFLY;